feat: re implement onMovementStep to be able to affect actor movement
This makes the function actually useful. It allows skills to stop an actor's movement at a particular tile based on conditions defined in the skill's `onMovementStep` function. Note that onMovementStep is called before moving to a tile. As long as all skills' onMovementStep returns true, the movement is valid. If even a single skill returns false, the movement is stopped at that tile prematurely. If any skill returns null, then the movement is stopped and the movement cost refunded.