feat: add a skill that automatically adds skills to other entities
This solves a problem whereby some skills require certain other "proccer" skills to be present on all other entities on the battlefield. The only solution to this is otherwise to add those dependency skills to all actors via a hook on actor.nut onInit. This is inefficient and unclean. This commit provides a clean and efficient way for the dependency skills to be added as necessary.